generalized date, time, image methods in MoodEditor class for CreateMoodActivity
